Fix the windows build when python is enabled.
This will only work with 64bit builds because the gtk in externals
is 64 bit. However, if 32 bit was needed it should be an easy fix.
This also depends on VS14 see lcm-1.0.0/lcm-python/setup.py where
that is set to use VS140COMNTOOLS. A future version of VS could
be used if the env var VS90COMNTOOLS was set to that versions
comntools env var.  This is also setup to work with Python2.7 64 bit.