Use dummy install logic instead of custom target
Use `install(CODE ...)` to add some no-op install logic in order to
trigger CMake to generate an install target. This is a little cleaner
than creating a custom target with a reserved name, and in particular
does not require fiddling with a CMake policy. (Critically, the policy
was missing a version check, and so was failing on older CMake.)